backend/requirements.txt
# Core
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic==2.*
pydantic-settings==2.*

# Database
sqlalchemy[asyncio]==2.*
asyncpg==0.30.*
alembic==1.14.*

# Cognee - GraphRAG knowledge engine
cognee==0.1.*

# Auth
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
python-multipart==0.0.*

# Storage (GCS)
google-cloud-storage==2.*

# Background tasks
celery[redis]==5.*
redis==5.*

# LLM providers
anthropic==0.*
google-generativeai==0.*

# Utils
python-dotenv==1.*
httpx==0.28.*
